Revert "Cache Patch"
