+++ main.py
@@ -1,5 +1,5 @@
 import os
-from src.bot import Bot
+from bot import Bot
 
 if __name__ == "__main__":
     music_bot = Bot(command_prefix='-')
